+++ R/biometryassist-deprecated.R
@@ -8,3 +8,4 @@
 #' @keywords internal
 #' @return No return value.
 NULL
+